requirements.txt
flask>=3.0
openai>=1.40
python-dotenv>=1.0
fpdf2>=2.8
